Refactor to shared logic
